+++ codebase/libs/radar/src/iwrf/IwrfCalib.cc
@@ -623,7 +623,7 @@ int IwrfCalib::readFromXmlFile(const string &calPath, string &errStr)
     TaStr::AddStr(errStr, "  ", strerror(errNum));
     return -1;
   }
-  int fileLen = calStat.st_size;
+  size_t fileLen = calStat.st_size;
 
   // open file
